V2H.CodeGenerator

Documentation

defBang :: Bang

moduleIdenToModuleInstanceIden :: ModuleIdentifierIR -> ModuleInstanceIdentifierIR

instanceHierarchyToString :: [ModuleInstanceIdentifierIR] -> [Char]

generateSumName :: [ModuleInstanceIdentifierIR] -> VariableOrNetIdentifierIR -> [Char]

generateDecConPair :: (Quote m, Lift t, HasField "identifier" p VariableOrNetIdentifierIR, HasField "dataType" p t) => [ModuleInstanceIdentifierIR] -> p -> m (Dec, Con)

generateSignalSumConstructors :: [ModuleInstanceIdentifierIR] -> [IR] -> IR -> Q [(Dec, Con)]

generateSignalSumTypesFromIRs :: [IR] -> IR -> ModuleInstanceIdentifierIR -> Q [Dec]

generateStringFromHierarchicalIdentifier :: HierarchicalIdentifierIR VariableOrNetIdentifierIR -> String

generateSignalIdentifier :: HierarchicalIdentifierIR VariableOrNetIdentifierIR -> Q Dec

generateSignalIdentifiers :: ExpandedIR -> Q [Dec]

Generate the hierarchial identifiers used by the second stage ir to reference signals

generateSignalType :: DataTypeIR -> Type

generateModuleName :: [ModuleInstanceIdentifierIR] -> [Char]

generateVariableFieldName :: VariableOrNetIdentifierIR -> [Char]

generateSubmoduleFieldName :: ModuleInstanceIdentifierIR -> [Char]

generateVariableSignal :: [ModuleInstanceIdentifierIR] -> VariableIR -> Type

generateSubmoduleFieldType :: [ModuleInstanceIdentifierIR] -> SubmoduleIR -> Type

generateSubmoduleNameFromHierarchicalIdentifier :: HierarchicalIdentifierIR a -> [Char]

generateCircuitRecord :: [ModuleInstanceIdentifierIR] -> [IR] -> IR -> Q [Dec]

Generate Static Representation of circuit for use by the EDSL

data X

Constructors

HVoni (HierarchicalIdentifierIR VariableOrNetIdentifierIR)

HDec (HierarchicalIdentifierIR (Name, Exp))

cutHierarchyTailInner :: Name -> (forall a. (Ord a, Show a, Eq a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> HierarchicalIdentifierIR a -> Exp -> X

cutHierarchyTail :: (Ord a, Show a, Eq a) => Name -> HierarchicalIdentifierIR a -> Exp -> X

cutHierarchyTailTop :: Name -> X -> Exp -> X

unwrap :: (forall x. HierarchicalIdentifierIR x -> t) -> X -> t

generateBody :: Name -> [X] -> Q Exp

generateConvertFromDynamicFunction :: Name -> ExpandedIR -> Q [Dec]

generateAccessorTail :: Exp -> HierarchicalIdentifierIR VariableOrNetIdentifierIR -> Exp

peel :: HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a

generateAccessor :: Name -> HierarchicalIdentifierIR VariableOrNetIdentifierIR -> Exp

generateMapElemStart :: Name -> HierarchicalIdentifierIR VariableOrNetIdentifierIR -> Q Exp

generateMapElemEnd :: Name -> HierarchicalIdentifierIR VariableOrNetIdentifierIR -> Q Exp

generateConvertToDynamicFunction :: Name -> ExpandedIR -> Q [Dec]

generateZeoredElem :: (Quote m, Lift t, Lift t, HasField "dataType" b t) => (t, b) -> m Exp

generateEmptyValue :: ExpandedIR -> Q [Dec]

generateExpandedIRValue :: (Quote m, Lift t, Lift t, Applicative f) => t -> t -> m (f Dec)

generateEval :: ExpandedIR -> Q [Dec]

generateInitState :: Q [Dec]