V2H.IRGenerator

Description

Module for converting Abstract Syntax Tree representation of circuit into First Stage IR and First Stage IR into Second Stage IR.

Documentation

generateConnection :: VariableLvalue -> ConnectionIR

generateStatementItemIRs :: VariableMapIR -> NetMapIR -> StatementItem -> StatementItemIR

Generate statementItemIRs from abstract syntax tree representation

generateEdgeIdentifier :: Maybe EdgeIdentifier -> Maybe EdgeIdentifierIR

generateAlwaysConstructIR :: VariableMapIR -> NetMapIR -> Int -> AlwaysConstruct -> (AlwaysConstructIdentifierIR, AlwaysConstructIR)

Generate always construct ir from abstract syntax tree representation

mapi :: (Num a, Enum a) => (a -> b -> b) -> [b] -> [b]

generateAlwaysConstructIRs :: VariableMapIR -> NetMapIR -> ModuleDeclaration -> AlwaysConstructMapIR

generateInnerVariableIRs :: [DataDeclaration] -> VariableMapIR

generatePortDirection :: PortDirection -> PortDirectionIR

generatePortVariableIR :: PortDeclaration -> Maybe (VariableOrNetIdentifierIR, VariableIR)

generatePortVariableIRs :: [PortDeclaration] -> VariableMapIR

generateVariableIRs :: (HasTypes p PortDeclaration, HasTypes p DataDeclaration) => p -> Map VariableOrNetIdentifierIR VariableIR

generateNetIRs :: p -> Map k a

generatePortDeclarationIR :: VariableMapIR -> NetMapIR -> PortDeclaration -> (PortIdentifierIR, PortDeclarationIR)

generatePortDeclarationIRs :: VariableMapIR -> NetMapIR -> ModuleDeclaration -> PortDeclarationMapIR

generateSubmoduleConnections :: VariableMapIR -> NetMapIR -> PortConnection -> (PortIdentifierIR, ExpressionIR)

generateSubmoduleIR :: VariableMapIR -> NetMapIR -> ModuleInstantiation -> SubmoduleIR

generateSubmoduleIRs :: VariableMapIR -> NetMapIR -> ModuleDeclaration -> [SubmoduleIR]

generateModuleIR :: ModuleDeclaration -> IR

generateIR :: SourceText -> Either String [IR]

Entry function to generate First Stage IR representation from abstract syntax tree.

unionIRExpanded :: ExpandedIR -> ExpandedIR -> ExpandedIR

connect :: SubmoduleIR -> (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> PortDeclarationIR -> Maybe (HierarchicalIdentifierIR VariableOrNetIdentifierIR, [(Maybe SelectIR, ConnectionIR)])

addIRsAndConnect :: [IR] -> IR -> (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> ExpandedIR -> SubmoduleIR -> ExpandedIR

updateConnectionIRHierarchicalIdentifierIRs :: (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> ConnectionIR -> ConnectionIR

updateExpressionIRHierarchicalIdentifiersIRs :: (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> ExpressionIR -> ExpressionIR

updateStatementItemIRHierarchicalIdentifierIRs :: (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> StatementItemIR -> StatementItemIR

updateEventExpressionIRHierarchicalIdentifierIR :: (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> EventExpressionIR -> EventExpressionIR

updateSensitivityIRHierarchicalIdentifierIRs :: (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> SensitivityIR -> SensitivityIR

updateAlwaysConstructIRHierarchicalIdentifierIRs :: (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> AlwaysConstructIR -> AlwaysConstructIR

generateExpandedIRSubmodule :: [IR] -> IR -> (forall a. (Ord a, Show a) => HierarchicalIdentifierIR a -> HierarchicalIdentifierIR a) -> ExpandedIR

Fill out hierarchy of modules in second stage IR. Uses rank-2 function type to wrap identifiers into hierarchical identifiers.

moduleIdenToModuleInstanceIden :: ModuleIdentifierIR -> ModuleInstanceIdentifierIR

generateExpandedIR :: IR -> [IR] -> ExpandedIR

Generate Second Stage IR representation from abstract first stage IR.