V2H.Simulator.Signal

Documentation

data SignalChange sig a

Constructors

SignalChange

Fields

_start :: a

_end :: a

mkStable :: forall {k} {a} {sig :: k}. a -> SignalChange sig a

mkStableFromSignalValue :: forall {k} {sig :: k} {a :: Nat}. SignalValue -> SignalChange sig (Signal a)

type family CheckBitSize (n :: Nat) (a :: Nat) :: Constraint where ...

Type family used to perform type level operations to check that signal literal fits in signal width

Equations

CheckBitSize n a = If (a <=? ((2 ^ n) - 1)) (() :: Constraint) (TypeError (Text "Literal too large"))

mkSignal :: forall a n. (CheckBitSize n a, KnownNat a) => Signal n

Proxy value used to check width of a signal literal using typefamily CheckBitSize

newtype Signal (a :: Nat)

Static representation of signals within the EDSL. Signal is parameterised by its width.

Constructors

Signal

Fields

value :: Integer

start :: forall k (sig :: k) a k (sig :: k). Lens (SignalChange (sig :: k) a) (SignalChange (sig :: k) a) a a

end :: forall k (sig :: k) a k (sig :: k). Lens (SignalChange (sig :: k) a) (SignalChange (sig :: k) a) a a