V2H.Simulator.Simulate

Documentation

data SignalValueChange

Data type used to represent a signal changing value

Constructors

SignalValueChange

Fields

connection :: ConnectionIR

old :: SignalValue

new :: SignalValue

fetchConnectionFromSignalValueChange :: SignalValueChange -> ConnectionIR

signalChangeToDynamicEnd :: (FetchHierarchicalIdentifierIR r, HasField "value" r Integer, HasField "_end" r r) => r -> SignalValue

signalChangeToDynamicStart :: (FetchHierarchicalIdentifierIR r, HasField "value" r Integer, HasField "_start" r r) => r -> SignalValue

class FetchHierarchicalIdentifierIR s where

Methods

fetchHierarchicalIdentifierIR :: s -> HierarchicalIdentifierIR VariableOrNetIdentifierIR

fetchSignalDataType :: s -> DataTypeIR

data StimulatedCircuit a

Data type manipulated by testbenches using EDSL

Constructors

StimulatedCircuit

Fields

_stimulatedState :: a

_stimulatedSignals :: Set (HierarchicalIdentifierIR VariableOrNetIdentifierIR)

stimulatedState :: forall a a. Lens (StimulatedCircuit a) (StimulatedCircuit a) a a

stimulatedSignals :: forall a. Lens' (StimulatedCircuit a) (Set (HierarchicalIdentifierIR VariableOrNetIdentifierIR))

mkStimulatedCircuit :: a -> StimulatedCircuit a

executeTimeSlot :: ExpandedIR -> State (DynamicCircuitState, TimeSlot) ()

Execute all events in a timeslot, execute a single event may generate more events. This function will only return once these extra events have been executed.

statementItemToTimeSlotEvent :: StatementItemIR -> [TimeSlotEvent]

Given a statement item from a process block, generate a time slot event

updateTimeSlotFromSensitiveProcess :: AlwaysConstructIR -> State TimeSlot ()

Apply updates events generated by a process block

doesTriggerLevelChange :: SignalValueChange -> ConnectionIR -> Bool

doesTriggerPosedge :: SignalValueChange -> ConnectionIR -> Bool

doesTriggerNegedge :: SignalValueChange -> ConnectionIR -> Bool

doesSignalChangeTriggerEvent :: SignalValueChange -> EventExpressionIR -> Bool

doesSignalChangeTriggerProcess :: SignalValueChange -> AlwaysConstructIR -> Bool

updateTimeSlotFromSignalValueUpdate :: Map (HierarchicalIdentifierIR VariableOrNetIdentifierIR) (Set AlwaysConstructIR) -> SignalValueChange -> State TimeSlot ()

generateTimeSlotEventWithIdentifier :: DynamicCircuitState -> HierarchicalIdentifierIR VariableOrNetIdentifierIR -> TimeSlotEvent

Given a stimulated circuit from the testbench, this function determines which update events to apply.

generateTimeSlotFromStimulatedCircuit :: ExpandedIR -> StimulatedCircuit DynamicCircuitState -> TimeSlot

Given a stimulated circuit from the testbench, determines which update events to apply.

getValueFromConnection :: DynamicCircuitState -> ConnectionIR -> SignalValue

Get value of a signal from a ConnectionIR

data ExpressionType

Constructors

SelfDetermined

ContextDetermined DataTypeIR

whenSD :: ExpressionType -> DataTypeIR -> DataTypeIR

chooseWith :: Ord a => (t -> a) -> (t, t) -> t

getDataTypeOfSelfDeterminedExpression :: ExpandedIR -> ExpressionIR -> DataTypeIR

Self-determined expressions within System Verilog define their result type. This function determines what this result type is.

evaluateExpression :: ExpandedIR -> ExpressionType -> DynamicCircuitState -> ExpressionIR -> SignalValue

Evaluate an expression within the System Verilog simulator

generateUpdateNetVariableFromConnection :: DynamicCircuitState -> HierarchicalIdentifierIR VariableOrNetIdentifierIR -> (Maybe SelectIR, ConnectionIR) -> TimeSlotEvent

propagateUpdate :: ExpandedIR -> ConnectionIR -> State (DynamicCircuitState, TimeSlot) ()

Given the change of a signal value, find all signals which depend on this value and update them too

executeEvent :: ExpandedIR -> TimeSlotEvent -> State (DynamicCircuitState, TimeSlot) ()

Execute Event and Update Circuit State and Timeslot. Executing an event may add many more events to the timeslot state. This circuit pattern matches against the type of event to execute.

executeEvents :: ExpandedIR -> Queue TimeSlotEvent -> State (DynamicCircuitState, TimeSlot) ()

executeActiveRegionSet :: ExpandedIR -> State (DynamicCircuitState, TimeSlot) ()

Execute all events in the active region set

executeReActiveRegionSet :: ExpandedIR -> State (DynamicCircuitState, TimeSlot) ()

Execute all events in the re-active region set

eval :: ExpandedIR -> StimulatedCircuit (DynamicCircuitState, DynamicCircuitState) -> DynamicCircuitState

Given the representation of the circuit and a list of circuit stimulations apply all updates to the circuit and return state of the circuit after all updates have been applied.

triggerAllAlwaysCombBlocks :: ExpandedIR -> DynamicCircuitState -> DynamicCircuitState

Trigger all combinatorial blocks. This is used to find the initial state of the circuit.