V2H.V2H

Documentation

safeReadFile :: FilePath -> IO (Either String ByteString)

setupWriteOutput :: FilePath -> String -> [FilePath] -> Q [Dec]

Debugging function to see what setup is generation

setup :: String -> [FilePath] -> Q [Dec]

Top level entry function for generating type information using V2H First argument is the toplevel module name of rtl design Second argument is the list to all filepaths used in the design